requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
pydantic
python-dotenv
langchain-text-splitters